update welcome page and rename Set-Finder to Set-Spotter
