--- Set-Finder/SceneDelegate.swift
@@ -1,79 +1,0 @@
-//
-//  SceneDelegate.swift
-//  Set-Finder
-//
-//  Created by JD del Alamo on 7/13/23.
-//
-
-import UIKit
-
-let SKIP_WELCOME_KEY = "SKIP_WELCOME"
-let HIGHLIGHT_SETS_KEY = "HIGHLIGHT_SETS"
-
-class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-
-   var window: UIWindow?
-
-
-   func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-      // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
-      // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
-      // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-      guard let windowScene = (scene as? UIWindowScene) else { return }
-
-      let skipWelcomeKeyPresent = UserDefaults.standard.object(forKey: SKIP_WELCOME_KEY) != nil
-      if (!skipWelcomeKeyPresent) {
-         // First time running app--set default to false
-         UserDefaults.standard.set(false, forKey: SKIP_WELCOME_KEY)
-      }
-      let skipWelcome = UserDefaults.standard.bool(forKey: SKIP_WELCOME_KEY)
-
-      let highlightSetsPresent = UserDefaults.standard.object(forKey: HIGHLIGHT_SETS_KEY) != nil
-      if (!highlightSetsPresent) {
-         // First time running app--set default to true
-         UserDefaults.standard.set(true, forKey: HIGHLIGHT_SETS_KEY)
-      }
-
-      let mainStoryboard = UIStoryboard(name: "Main", bundle: nil)
-      var rootViewController: UIViewController
-      if (skipWelcome) {
-         rootViewController = mainStoryboard.instantiateViewController(withIdentifier: "main") as UIViewController
-      } else {
-         rootViewController = mainStoryboard.instantiateViewController(withIdentifier: "welcome") as UIViewController
-      }
-
-      let window = UIWindow(windowScene: windowScene)
-      window.rootViewController = rootViewController
-      self.window = window
-      self.window?.makeKeyAndVisible()
-   }
-
-   func sceneDidDisconnect(_ scene: UIScene) {
-      // Called as the scene is being released by the system.
-      // This occurs shortly after the scene enters the background, or when its session is discarded.
-      // Release any resources associated with this scene that can be re-created the next time the scene connects.
-      // The scene may re-connect later, as its session was not necessarily discarded (see `application:didDiscardSceneSessions` instead).
-   }
-
-   func sceneDidBecomeActive(_ scene: UIScene) {
-      // Called when the scene has moved from an inactive state to an active state.
-      // Use this method to restart any tasks that were paused (or not yet started) when the scene was inactive.
-   }
-
-   func sceneWillResignActive(_ scene: UIScene) {
-      // Called when the scene will move from an active state to an inactive state.
-      // This may occur due to temporary interruptions (ex. an incoming phone call).
-   }
-
-   func sceneWillEnterForeground(_ scene: UIScene) {
-      // Called as the scene transitions from the background to the foreground.
-      // Use this method to undo the changes made on entering the background.
-   }
-
-   func sceneDidEnterBackground(_ scene: UIScene) {
-      // Called as the scene transitions from the foreground to the background.
-      // Use this method to save data, release shared resources, and store enough scene-specific state information
-      // to restore the scene back to its current state.
-   }
-
-}
