+++ Set-Spotter.xcodeproj/xcuserdata/jd.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "412F8CEA-A272-419D-A58A-FB9D181DDF67"
+   type = "1"
+   version = "2.0">
+</Bucket>
