+++ Set-Spotter/Assets.xcassets/AccentColor.colorset/Contents.json
