+++ Set-Spotter/Assets.xcassets/Contents.json
