+++ Set-Spotter/Set-Spotter-Bridging-Header.h
